+++ eabm/eabm_package/python/single_element_compression_mesh_convergence.py
@@ -1,0 +1,3 @@
+parameter_schema = {
+    'global_seed': [1.0, 0.5, 0.25, 0.125],
+}