+++ eabm/eabm_package/python/single_element_compression_sobol_sequence.py
@@ -1,5 +1,13 @@
 parameter_schema = {
     'num_simulations': 4,
-    'width': [0.9, 1.1],
-    'height': [0.9, 1.1]
+    'width': {
+        'distribution': 'uniform',
+        'loc': 0.9,
+        'scale': 0.2,
+    },
+    'height': {
+        'distribution': 'uniform',
+        'loc': 0.9,
+        'scale': 0.2,
+    },
 }